Rename kbun to kbname
