+++ go/kbname/username.go
@@ -1,7 +1,7 @@
 // Copyright 2018 Keybase, Inc. All rights reserved. Use of
 // this source code is governed by the included BSD license.
 
-package kbun
+package kbname
 
 import (
 	"regexp"
